moved validate domain to parser package
