+++ parser/parser.go
@@ -1,7 +1,9 @@
-package spf
+package parser
 
 import (
+	"errors"
 	"fmt"
+	"golang.org/x/net/idna"
 	"net"
 	"strconv"
 	"strings"
@@ -48,6 +50,16 @@ type Record struct {
 	Exp      *Modifier
 	Unknown  []Modifier
 }
+
+// Errors returned by ValidateDomain.  Each corresponds to one of the
+// syntax checks described in RFC 7208 section 4.3.
+var (
+	ErrSingleLabel    = errors.New("domain must have at least two labels")
+	ErrEmptyLabel     = errors.New("domain has empty label")
+	ErrLabelTooLong   = errors.New("domain label exceeds 63 octets")
+	ErrDomainTooLong  = errors.New("domain exceeds 255 octets")
+	ErrIDNAConversion = errors.New("IDNA ToASCII failed")
+)
 
 /* ========= public parser entry-point ========= */
 // Parse checks the record syntax defined in RFC 7208 section 4.6 and returns a structured representation.
@@ -425,4 +437,61 @@ func parseExists(q Qualifier, rest string) (*Mechanism, error) {
 		Domain: spec, // raw, possibly macro-containing string
 		Macro:  hasMacro,
 	}, nil
+}
+
+// ValidateDomain normalises and validates a raw domain name, according to
+// RFC 7208, section 4.3.
+//
+// Validation steps:
+//
+//  1. Remove one trailing dot because domains are implicitly absolute.
+//
+//  2. Convert the name to its Punycode A-label form with idna.Lookup.ToASCII.
+//
+//  3. Apply SPF pre-evaluation checks:
+//
+//     * Overall length must not exceed 255 octets.
+//     * The domain must contain at least two labels (must include a dot).
+//     * No empty label may appear except the implicit root.
+//     * Each label must be 1–63 octets long.
+//     * Labels may contain only lower-case letters, digits, and hyphens.
+//     * A hyphen may not appear at the start or end of any label.
+//
+// On success the function returns the ASCII (lower-case) domain and nil.
+// On failure, it returns an empty string along with a sentinel error.
+func ValidateDomain(raw string) (string, error) {
+	raw = strings.TrimSpace(raw)
+	// Trim the single trailing dot if any
+	raw = strings.TrimSuffix(raw, ".")
+
+	// convert to A-label RFC 5890 section 2.3
+	ascii, err := idna.Lookup.ToASCII(raw)
+	if err != nil {
+		return "", ErrIDNAConversion
+	}
+	ascii = strings.ToLower(ascii)
+
+	// check overall length limit
+	if len(ascii) > 255 {
+		return "", ErrDomainTooLong
+	}
+
+	labels := strings.Split(ascii, ".")
+	if len(labels) < 2 {
+		return "", ErrSingleLabel
+	}
+
+	for _, lbl := range labels {
+		switch {
+		case len(lbl) == 0:
+			return "", ErrEmptyLabel
+
+		case len(lbl) > 63:
+			return "", ErrLabelTooLong
+
+		}
+
+	}
+
+	return ascii, nil
 }
